+++ tests/MoneyTest.php
@@ -185,6 +185,8 @@ final class MoneyTest extends \PHPUnit_Framework_TestCase
             [self::AMOUNT + 1, new Currency(self::CURRENCY), false],
             [self::AMOUNT, new Currency(self::OTHER_CURRENCY), false],
             [self::AMOUNT + 1, new Currency(self::OTHER_CURRENCY), false],
+            [(string) self::AMOUNT, new Currency(self::CURRENCY), true],
+            [((string) self::AMOUNT).'.000', new Currency(self::CURRENCY), true],
         ];
     }