+++ tests/NumberTest.php
@@ -14,13 +14,13 @@ final class NumberTest extends \PHPUnit_Framework_TestCase
     {
         $number = Number::fromString($number);
 
-        $this->assertEquals($decimal, $number->isDecimal());
-        $this->assertEquals($half, $number->isHalf());
-        $this->assertEquals($currentEven, $number->isCurrentEven());
-        $this->assertEquals($negative, $number->isNegative());
-        $this->assertEquals($integerPart, $number->getIntegerPart());
-        $this->assertEquals($fractionalPart, $number->getFractionalPart());
-        $this->assertEquals($negative ? '-1' : '1', $number->getIntegerRoundingMultiplier());
+        $this->assertSame($decimal, $number->isDecimal());
+        $this->assertSame($half, $number->isHalf());
+        $this->assertSame($currentEven, $number->isCurrentEven());
+        $this->assertSame($negative, $number->isNegative());
+        $this->assertSame($integerPart, $number->getIntegerPart());
+        $this->assertSame($fractionalPart, $number->getFractionalPart());
+        $this->assertSame($negative ? '-1' : '1', $number->getIntegerRoundingMultiplier());
     }
 
     public function numberExamples()
@@ -44,7 +44,9 @@ final class NumberTest extends \PHPUnit_Framework_TestCase
             ['10.500', true, true, true, false, '10', '5'],
             ['10.9', true, false, true, false, '10', '9'],
             ['-10', false, false, true, true, '-10', ''],
+            ['-0', false, false, true, true, '-0', ''],
             ['-10.5', true, true, true, true, '-10', '5'],
+            ['-.5', true, true, true, true, '-0', '5'],
             ['.5', true, true, true, false, '0', '5'],
             [(string) PHP_INT_MAX, false, false, false, false, (string) PHP_INT_MAX, ''],
             [(string) -PHP_INT_MAX, false, false, false, true, (string) -PHP_INT_MAX, ''],
@@ -75,6 +77,31 @@ final class NumberTest extends \PHPUnit_Framework_TestCase
                 substr(PHP_INT_MAX, 0, strlen((string) PHP_INT_MAX) - 1).str_repeat('0', strlen((string) PHP_INT_MAX) - 1).PHP_INT_MAX,
                 '',
             ],
+        ];
+    }
+
+    /**
+     * @dataProvider invalidNumberExamples
+     * @expectedException \InvalidArgumentException
+     * @test
+     */
+    public function it_fails_parsing_invalid_numbers($number)
+    {
+        Number::fromString($number);
+    }
+
+    public function invalidNumberExamples()
+    {
+        return [
+            [''],
+            ['000'],
+            ['005'],
+            ['123456789012345678-123456'],
+            ['---123'],
+            ['123456789012345678+13456'],
+            ['-123456789012345678.-13456'],
+            ['+123456789'],
+            ['+123456789012345678.+13456'],
         ];
     }
 }